requirements.txt
lightning~=2.3.3
numpy~=1.26.4
torch~=2.4.1
matplotlib~=3.9.3
seaborn~=0.13.2
omegaconf~=2.3.0
SLRT~=0.1.0
typing_extensions~=4.12.2
scipy~=1.14.1
tensorflow~=2.17.0
transformers~=4.47.0
hydra-core~=1.3.2
psutil~=5.9.0
wandb~=0.19.0
opencv-python~=4.10.0.84
h5py~=3.12.1
pandas~=2.2.3
torchvision~=0.19.1
tqdm~=4.66.6
pillow~=11.0.0
gradio~=5.8.0
setuptools~=75.1.0